+++ core/src/storage/command.rs
@@ -5,6 +5,7 @@ use async_trait::async_trait;
 
 #[derive(Clone, Debug, PartialEq)]
 pub enum StorageCommand {
+    Delete { name: String },
     Journal,
     Load { name: String },
     Save { name: String },
@@ -13,7 +14,14 @@ pub enum StorageCommand {
 impl StorageCommand {
     fn summarize(&self, thing: Option<&Thing>) -> String {
         if let Some(thing) = thing {
+            let thing_type = match thing {
+                Thing::Location(_) => "location",
+                Thing::Npc(_) => "NPC",
+                Thing::Region(_) => "region",
+            };
+
             match self {
+                Self::Delete { .. } => format!("remove {} from journal", thing_type),
                 Self::Journal { .. } => unreachable!(),
                 Self::Load { .. } => {
                     if thing.uuid().is_some() {
@@ -22,19 +30,11 @@ impl StorageCommand {
                         format!("{} (unsaved)", thing.display_description())
                     }
                 }
-                Self::Save { .. } => {
-                    format!(
-                        "save {} to journal",
-                        match thing {
-                            Thing::Location(_) => "location",
-                            Thing::Npc(_) => "NPC",
-                            Thing::Region(_) => "region",
-                        }
-                    )
-                }
+                Self::Save { .. } => format!("save {} to journal", thing_type),
             }
         } else {
             match self {
+                Self::Delete { .. } => "remove an entry from journal",
                 Self::Journal { .. } => "list journal contents",
                 Self::Load { .. } => "load an entry",
                 Self::Save { .. } => "save an entry to journal",
@@ -91,6 +91,13 @@ impl Runnable for StorageCommand {
 
                 Ok(output)
             }
+            Self::Delete { name } => {
+                if !app_meta.data_store_enabled {
+                    return Err("The journal is not supported by your browser.".to_string());
+                }
+
+                repository::delete_by_name(app_meta, name).await
+            }
             Self::Load { name } => {
                 let thing = repository::load(app_meta, name);
                 let mut save_command = None;
@@ -133,6 +140,10 @@ impl Runnable for StorageCommand {
             vec![Self::Load {
                 name: input.to_string(),
             }]
+        } else if let Some(name) = input.strip_prefix("delete ") {
+            vec![Self::Delete {
+                name: name.to_string(),
+            }]
         } else if let Some(name) = input.strip_prefix("load ") {
             vec![Self::Load {
                 name: name.to_string(),
@@ -156,7 +167,7 @@ impl Runnable for StorageCommand {
 
         let mut result = Vec::new();
 
-        ["load", "save"]
+        ["delete", "load", "save"]
             .iter()
             .filter(|s| s.starts_with(input))
             .filter_map(|s| {
@@ -179,7 +190,7 @@ impl Runnable for StorageCommand {
             )
             .for_each(|(s, command)| result.push((s, command.summarize(None))));
 
-        let (input_prefix, input_name) = if let Some(parts) = ["load ", "save "]
+        let (input_prefix, input_name) = if let Some(parts) = ["delete ", "load ", "save "]
             .iter()
             .find_map(|prefix| input.strip_prefix(prefix).map(|name| (*prefix, name)))
         {
@@ -198,10 +209,9 @@ impl Runnable for StorageCommand {
                     .value()
                     .map(|name| {
                         if name.starts_with(input_name) {
-                            if input_prefix == "save " && thing.uuid().is_some() {
-                                None
-                            } else {
-                                Some((input_prefix, thing))
+                            match (input_prefix, thing.uuid()) {
+                                ("save ", Some(_)) | ("delete ", None) => None,
+                                _ => Some((input_prefix, thing)),
                             }
                         } else if name.starts_with(input) {
                             Some(("", thing))
@@ -448,6 +458,17 @@ mod test {
         );
 
         assert_eq!(
+            [(
+                "delete Potato & Potato, Esq.",
+                "remove location from journal"
+            )]
+            .iter()
+            .map(|(a, b)| (a.to_string(), b.to_string()))
+            .collect::<Vec<_>>(),
+            StorageCommand::autocomplete("delete P", &app_meta),
+        );
+
+        assert_eq!(
             [
                 ("save Potato Johnson", "save NPC to journal"),
                 ("save potato should be capitalized", "save NPC to journal"),
@@ -468,6 +489,14 @@ mod test {
             .map(|(a, b)| (a.to_string(), b.to_string()))
             .collect::<Vec<_>>(),
             StorageCommand::autocomplete("load P", &app_meta),
+        );
+
+        assert_eq!(
+            [("delete [name]", "remove an entry from journal")]
+                .iter()
+                .map(|(a, b)| (a.to_string(), b.to_string()))
+                .collect::<Vec<_>>(),
+            StorageCommand::autocomplete("delete", &app_meta),
         );
 
         assert_eq!(